Select support functions

When creating a Darwin Core archive, it is often useful to select only those fields that conform to the standard. These functions provide a vector of terms that can be used in combination with dplyr::select() and dplyr::any_of() to quickly select Darwin Core terms for the relevant data type (events, occurrences, media).

Usage

occurrence_terms()

event_terms()

Value

A vector of accepted (but not mandatory) values for that use case.

See also

Examples

# Return a vector of accepted terms in an Occurrence-based dataset
occurrence_terms()
#>   [1] "type"                               
#>   [2] "modified"                           
#>   [3] "language"                           
#>   [4] "license"                            
#>   [5] "rightsHolder"                       
#>   [6] "accessRights"                       
#>   [7] "bibliographicCitation"              
#>   [8] "references"                         
#>   [9] "institutionID"                      
#>  [10] "collectionID"                       
#>  [11] "datasetID"                          
#>  [12] "institutionCode"                    
#>  [13] "collectionCode"                     
#>  [14] "datasetName"                        
#>  [15] "ownerInstitutionCode"               
#>  [16] "basisOfRecord"                      
#>  [17] "informationWithheld"                
#>  [18] "dataGeneralizations"                
#>  [19] "dynamicProperties"                  
#>  [20] "occurrenceID"                       
#>  [21] "catalogNumber"                      
#>  [22] "recordNumber"                       
#>  [23] "recordedBy"                         
#>  [24] "recordedByID"                       
#>  [25] "individualCount"                    
#>  [26] "organismQuantity"                   
#>  [27] "organismQuantityType"               
#>  [28] "sex"                                
#>  [29] "lifeStage"                          
#>  [30] "reproductiveCondition"              
#>  [31] "caste"                              
#>  [32] "behavior"                           
#>  [33] "vitality"                           
#>  [34] "establishmentMeans"                 
#>  [35] "degreeOfEstablishment"              
#>  [36] "pathway"                            
#>  [37] "georeferenceVerificationStatus"     
#>  [38] "occurrenceStatus"                   
#>  [39] "associatedMedia"                    
#>  [40] "associatedOccurrences"              
#>  [41] "associatedReferences"               
#>  [42] "associatedTaxa"                     
#>  [43] "otherCatalogNumbers"                
#>  [44] "occurrenceRemarks"                  
#>  [45] "organismID"                         
#>  [46] "organismName"                       
#>  [47] "organismScope"                      
#>  [48] "associatedOrganisms"                
#>  [49] "previousIdentifications"            
#>  [50] "organismRemarks"                    
#>  [51] "materialEntityID"                   
#>  [52] "preparations"                       
#>  [53] "disposition"                        
#>  [54] "verbatimLabel"                      
#>  [55] "associatedSequences"                
#>  [56] "materialEntityRemarks"              
#>  [57] "materialSampleID"                   
#>  [58] "eventID"                            
#>  [59] "parentEventID"                      
#>  [60] "eventType"                          
#>  [61] "fieldNumber"                        
#>  [62] "eventDate"                          
#>  [63] "eventTime"                          
#>  [64] "startDayOfYear"                     
#>  [65] "endDayOfYear"                       
#>  [66] "year"                               
#>  [67] "month"                              
#>  [68] "day"                                
#>  [69] "verbatimEventDate"                  
#>  [70] "habitat"                            
#>  [71] "samplingProtocol"                   
#>  [72] "sampleSizeValue"                    
#>  [73] "sampleSizeUnit"                     
#>  [74] "samplingEffort"                     
#>  [75] "fieldNotes"                         
#>  [76] "eventRemarks"                       
#>  [77] "locationID"                         
#>  [78] "higherGeographyID"                  
#>  [79] "higherGeography"                    
#>  [80] "continent"                          
#>  [81] "waterBody"                          
#>  [82] "islandGroup"                        
#>  [83] "island"                             
#>  [84] "country"                            
#>  [85] "countryCode"                        
#>  [86] "stateProvince"                      
#>  [87] "county"                             
#>  [88] "municipality"                       
#>  [89] "locality"                           
#>  [90] "verbatimLocality"                   
#>  [91] "minimumElevationInMeters"           
#>  [92] "maximumElevationInMeters"           
#>  [93] "verbatimElevation"                  
#>  [94] "verticalDatum"                      
#>  [95] "minimumDepthInMeters"               
#>  [96] "maximumDepthInMeters"               
#>  [97] "verbatimDepth"                      
#>  [98] "minimumDistanceAboveSurfaceInMeters"
#>  [99] "maximumDistanceAboveSurfaceInMeters"
#> [100] "locationAccordingTo"                
#> [101] "locationRemarks"                    
#> [102] "decimalLatitude"                    
#> [103] "decimalLongitude"                   
#> [104] "geodeticDatum"                      
#> [105] "coordinateUncertaintyInMeters"      
#> [106] "coordinatePrecision"                
#> [107] "pointRadiusSpatialFit"              
#> [108] "verbatimCoordinates"                
#> [109] "verbatimLatitude"                   
#> [110] "verbatimLongitude"                  
#> [111] "verbatimCoordinateSystem"           
#> [112] "verbatimSRS"                        
#> [113] "footprintWKT"                       
#> [114] "footprintSRS"                       
#> [115] "footprintSpatialFit"                
#> [116] "georeferencedBy"                    
#> [117] "georeferencedDate"                  
#> [118] "georeferenceProtocol"               
#> [119] "georeferenceSources"                
#> [120] "georeferenceRemarks"                
#> [121] "geologicalContextID"                
#> [122] "earliestEonOrLowestEonothem"        
#> [123] "latestEonOrHighestEonothem"         
#> [124] "earliestEraOrLowestErathem"         
#> [125] "latestEraOrHighestErathem"          
#> [126] "earliestPeriodOrLowestSystem"       
#> [127] "latestPeriodOrHighestSystem"        
#> [128] "earliestEpochOrLowestSeries"        
#> [129] "latestEpochOrHighestSeries"         
#> [130] "earliestAgeOrLowestStage"           
#> [131] "latestAgeOrHighestStage"            
#> [132] "lowestBiostratigraphicZone"         
#> [133] "highestBiostratigraphicZone"        
#> [134] "lithostratigraphicTerms"            
#> [135] "group"                              
#> [136] "formation"                          
#> [137] "member"                             
#> [138] "bed"                                
#> [139] "identificationID"                   
#> [140] "verbatimIdentification"             
#> [141] "identificationQualifier"            
#> [142] "typeStatus"                         
#> [143] "identifiedBy"                       
#> [144] "identifiedByID"                     
#> [145] "dateIdentified"                     
#> [146] "identificationReferences"           
#> [147] "identificationVerificationStatus"   
#> [148] "identificationRemarks"              
#> [149] "taxonID"                            
#> [150] "scientificNameID"                   
#> [151] "acceptedNameUsageID"                
#> [152] "parentNameUsageID"                  
#> [153] "originalNameUsageID"                
#> [154] "nameAccordingToID"                  
#> [155] "namePublishedInID"                  
#> [156] "taxonConceptID"                     
#> [157] "scientificName"                     
#> [158] "acceptedNameUsage"                  
#> [159] "parentNameUsage"                    
#> [160] "originalNameUsage"                  
#> [161] "nameAccordingTo"                    
#> [162] "namePublishedIn"                    
#> [163] "namePublishedInYear"                
#> [164] "higherClassification"               
#> [165] "kingdom"                            
#> [166] "phylum"                             
#> [167] "class"                              
#> [168] "order"                              
#> [169] "superfamily"                        
#> [170] "family"                             
#> [171] "subfamily"                          
#> [172] "tribe"                              
#> [173] "subtribe"                           
#> [174] "genus"                              
#> [175] "genericName"                        
#> [176] "subgenus"                           
#> [177] "infragenericEpithet"                
#> [178] "specificEpithet"                    
#> [179] "infraspecificEpithet"               
#> [180] "cultivarEpithet"                    
#> [181] "taxonRank"                          
#> [182] "verbatimTaxonRank"                  
#> [183] "scientificNameAuthorship"           
#> [184] "vernacularName"                     
#> [185] "nomenclaturalCode"                  
#> [186] "taxonomicStatus"                    
#> [187] "nomenclaturalStatus"                
#> [188] "taxonRemarks"                       
#> [189] "measurementID"                      
#> [190] "parentMeasurementID"                
#> [191] "measurementType"                    
#> [192] "measurementValue"                   
#> [193] "measurementAccuracy"                
#> [194] "measurementUnit"                    
#> [195] "measurementDeterminedBy"            
#> [196] "measurementDeterminedDate"          
#> [197] "measurementMethod"                  
#> [198] "measurementRemarks"                 
#> [199] "resourceRelationshipID"             
#> [200] "resourceID"                         
#> [201] "relationshipOfResourceID"           
#> [202] "relatedResourceID"                  
#> [203] "relationshipOfResource"             
#> [204] "relationshipAccordingTo"            
#> [205] "relationshipEstablishedDate"        
#> [206] "relationshipRemarks"                

# Use this vector to filter a data frame
df <- tibble::tibble(
  name = c("Crinia Signifera", "Crinia Signifera", "Litoria peronii"),
  longitude = c(35.27, 35.24, 35.83),
  latitude = c(149.33, 149.34, 149.34),
  eventDate = c("2010-10-14", "2010-10-14", "2010-10-14"),
  measurement1 = c(24.3, 24.9, 20.1), # example measurement column
  measurement2 = c(0.92, 1.03, 1.09)  # example measurement column
  )

df |>
  select(any_of(occurrence_terms()))
#> Error in select(df, any_of(occurrence_terms())): could not find function "select"