Dataset of supported Darwin Core terms

The Darwin Core Standard is maintained by Biodiversity Information Standards, previously known as the Taxonomic Databases Working Group and known by the acronym 'TDWG'. This tibble is the full list of supported terms, current as at 2024-12-10.

Usage

darwin_core_terms

Format

A tibble with 206 rows and 6 columns:

class

TDWG group that a term belongs to.

term

Column header names that can be used in Darwin Core

url

Stable url to information describing the term.

definition

Human-readable definition of the term.

comments

Further information from TDWG.

examples

Examples of how the field should be populated.

Source

Slightly modified version of a table supplied by TDWG at this link.

See also