Convert columns with measurement data for an individual or event to Darwin Core standard

This function is a work in progress, and should be used with caution.

In raw collected data, many types of information can be captured in one column. For example, the column name LMA_g.m2 contains the measured trait (Leaf Mass per Area, LMA) and the unit of measurement (grams per meter squared, g/m2), and recorded in that column are the values themselves. In Darwin Core, these different types of information must be separated into multiple columns so that they can be ingested correctly and aggregated with sources of data accurately.

This function converts information preserved in a single measurement column into multiple columns (measurementID, measurementUnit, and measurementType) as per Darwin Core standard.

Usage

set_measurements(.df, cols = NULL, unit = NULL, type = NULL, .keep = "unused")

Arguments

.df

a data.frame or tibble that the column should be appended to.

cols

vector of column names to be included as 'measurements'. Unquoted.

unit

vector of strings giving units for each variable

type

vector of strings giving a description for each variable

.keep

Control which columns from .data are retained in the output. Note that unlike dplyr::mutate(), which defaults to "all" this defaults to "unused"; i.e. only keeps Darwin Core fields, and not those fields used to generate them.

Value

A tibble with the requested fields added.

Details

Columns are nested in a single column measurementOrFact that contains Darwin Core standard measurement fields. By nesting three measurement columns within the measurementOrFact, despite measurement columns converting to long format (one row per measurement, per occurrence), data will remain organised by occurrences (one row per occurrences). Data can be unnested into long format using dplyr::unnest()