Add information of individual organisms to a tibble

Format fields that contain measurements or attributes of individual organisms to a tibble. Fields include those that specify sex, life stage or condition. Individuals can be identified by an individualID if data contains resampling.

In practice this is no different from using mutate(), but gives some informative errors, and serves as a useful lookup for fields in the Darwin Core Standard.

Usage

use_individual_traits(
  .df,
  individualID = NULL,
  lifeStage = NULL,
  sex = NULL,
  vitality = NULL,
  reproductiveCondition = NULL,
  .keep = "unused"
)

Arguments

.df

a data.frame or tibble that the column should be appended to.

individualID

An identifier for an individual or named group of individual organisms represented in the Occurrence. Meant to accommodate resampling of the same individual or group for monitoring purposes. May be a global unique identifier or an identifier specific to a data set.

lifeStage

The age class or life stage of an organism at the time of occurrence.

sex

The sex of the biological individual.

vitality

An indication of whether an organism was alive or dead at the time of collection or observation.

reproductiveCondition

The reproductive condition of the biological individual.

.keep

Control which columns from .data are retained in the output. Note that unlike dplyr::mutate(), which defaults to "all" this defaults to "unused"; i.e. only keeps Darwin Core fields, and not those fields used to generate them.

Value

A tibble with the requested fields added.

Details

Examples of lifeStage values:

zygote

larva

adult

seedling

flowering

Examples of vitality values:

alive

dead

uncertain

Examples of reproductiveCondition values:

non-reproductive

pregnant

in bloom

fruit bearing

See also