Add locality data to a tibble

Locality information refers to a description of a place, rather than a spatial coordinate.

Usage

use_locality(
  .df,
  continent = NULL,
  country = NULL,
  countryCode = NULL,
  stateProvince = NULL,
  locality = NULL,
  .keep = "unused"
)

Arguments

.df

a data.frame or tibble that the column should be appended to.

continent

(string) Valid continent. See details.

country

Valid country name. See country_codes.

countryCode

Valid country code. See country_codes.

stateProvince

A sub-national region.

locality

A specific location, such as a property or address.

.keep

Control which columns from .data are retained in the output. Note that unlike dplyr::mutate(), which defaults to "all" this defaults to "unused"; i.e. only keeps Darwin Core fields, and not those fields used to generate them.

Value

A tibble with the requested fields.

Details

Example values are:

continent should be one of "Africa", "Antarctica", "Asia", "Europe", "North America", "Oceania" or "South America".

countryCode should be supplied according to the ISO 3166-1 ALPHA-2 standard, as per TDWG advice.

See also