Add spatial fields to a tibble using sf POINT coordinates

This function helps format standard location fields to a tibble. It differs from use_coordinates() by accepting sf geometry columns of class POINT as coordinates (rather than numeric lat/lon coordinates). The advantage of using an sf geometry is that the Coordinate Reference System (CRS) is automatically formatted into the required geodeticDatum column.

Usage

use_sf(.df, geometry = NULL, .keep = "unused")

Arguments

.df

a data.frame or tibble that the column should be appended to.

geometry

the latitude/longitude coordinates in decimal degrees as sf POINT class

.keep

Control which columns from .data are retained in the output. Note that unlike dplyr::mutate(), which defaults to "all" this defaults to "unused"; i.e. only keeps Darwin Core fields, and not those fields used to generate them.

Value

A tibble with the requested fields added.

See also