Add taxonomic information to a tibble

Format fields that contain taxonomic name information from kingdom to species, as well as the common/vernacular name, to a tibble.

Usage

use_taxonomy(
  .df,
  kingdom = NULL,
  phylum = NULL,
  class = NULL,
  order = NULL,
  family = NULL,
  genus = NULL,
  specificEpithet = NULL,
  vernacularName = NULL,
  .keep = "unused"
)

Arguments

kingdom

The kingdom name of identified taxon.

phylum

The phylum name of identified taxon.

class

The class name of identified taxon.

order

The order name of identified taxon.

family

The family name of identified taxon.

genus

The genus name of the identified taxon.

specificEpithet

The name of the first species or species epithet of the scientificName. See documentation

vernacularName

The common or vernacular name of the identified taxon.

.keep

Control which columns from .data are retained in the output. Note that unlike dplyr::mutate, which defaults to "all" this defaults to "unused"; i.e. only keeps Darwin Core fields, and not those fields used to generate them.

df

a data.frame or tibble that the column should be appended to.

Value

A tibble with the requested fields added.

Details

In practice this is no different from using mutate(), but gives some informative errors, and serves as a useful lookup for taxonomic names in the Darwin Core Standard.

Examples of specificEphithet:

If scientificName is Abies concolor, the specificEpithet is concolor.

If scientificName is Semisulcospira gottschei, the specificEpithet is gottschei.

See also