Create identifier columns

Identifiers are columns that uniquely identify a single record within a dataset. These are helper functions, designed to make it easier to generate such columns from a given dataset. They are designed to be called within use_events(), use_occurrences(), or (equivalently) dplyr::mutate(). Generally speaking, it is best practice to use existing information from a dataset to generate identifiers; for this reason we recomment using composite_id() to aggregate existing fields, if no such composite is already present within the dataset. It is possible to call sequential_id() or random_id() within composite_id() to combine existing and new columns.

Usage

composite_id(..., sep = "-")

sequential_id(width)

random_id()

Arguments

...

Zero or more variable names from the tibble being mutated (unquoted), and/or zero or more _id functions, separated by commas.

sep

Character used to separate field values. Defaults to "-"

width

(Integer) how many characters should the resulting string be? Defaults to one plus the order of magnitude of the largest number.

Value

An amended tibble, containing a field with the requested information.

Examples

library(tibble)
df <- tibble(eventDate = paste0(rep(c(2020:2024), 3), "-01-01"),
             basisOfRecord = "humanObservation",
             site = rep(c("A01", "A02", "A03"), each = 5))
df |>
    use_occurrences(occurrenceID = composite_id(sequential_id(),
                                                site,
                                                eventDate))
#> ⠙ Checking 2 columns: basisOfRecord and occurrenceID
#> ✔ Checking 2 columns: basisOfRecord and occurrenceID [626ms]
#> 
#> # A tibble: 15 × 4
#>    eventDate  basisOfRecord    site  occurrenceID      
#>    <chr>      <chr>            <chr> <chr>             
#>  1 2020-01-01 humanObservation A01   001-A01-2020-01-01
#>  2 2021-01-01 humanObservation A01   002-A01-2021-01-01
#>  3 2022-01-01 humanObservation A01   003-A01-2022-01-01
#>  4 2023-01-01 humanObservation A01   004-A01-2023-01-01
#>  5 2024-01-01 humanObservation A01   005-A01-2024-01-01
#>  6 2020-01-01 humanObservation A02   006-A02-2020-01-01
#>  7 2021-01-01 humanObservation A02   007-A02-2021-01-01
#>  8 2022-01-01 humanObservation A02   008-A02-2022-01-01
#>  9 2023-01-01 humanObservation A02   009-A02-2023-01-01
#> 10 2024-01-01 humanObservation A02   010-A02-2024-01-01
#> 11 2020-01-01 humanObservation A03   011-A03-2020-01-01
#> 12 2021-01-01 humanObservation A03   012-A03-2021-01-01
#> 13 2022-01-01 humanObservation A03   013-A03-2022-01-01
#> 14 2023-01-01 humanObservation A03   014-A03-2023-01-01
#> 15 2024-01-01 humanObservation A03   015-A03-2024-01-01